small changes cleanup; moved some stuff to nebu; some big changes I already forgot about; merge of branches/drafting

+++ nebu/include/base/nebu_types.h
@@ -11,7 +11,7 @@ typedef struct {
 } SystemTime;
 
 typedef struct {
-	float x, y, width, height;
+	int x, y, width, height;
 } nebu_Rect;
 
 #endif
